+++ client/src/components/dummyDataUsers.js
@@ -1,0 +1,16 @@
+const users = [
+    {
+        username: "Vincent Porter",
+        icon: "https://bootdey.com/img/Content/avatar/avatar1.png"
+    },
+    {
+        username: "Aiden Chavez",
+        icon: "https://bootdey.com/img/Content/avatar/avatar2.png"
+    },
+    {
+        username: "Mike Thomas",
+        icon: "https://bootdey.com/img/Content/avatar/avatar3.png"
+    }
+]
+
+export default users;
